Allow `invokespecial` and `invokestatic` to refer to interface methods.
Previously the parser would only work if `invokespecial` or
`invokestatic` referred to an ordinary method, and would fail if it
found a reference to an interface method. But according to section
6.5 of the JVM spec, "the run-time constant pool item ... must be a
symbolic reference to a method or an interface method." This patch
updates the parser to comply with the spec.