Fix -Wincomplete-uni-patterns warnings
GHC 9.2 now includes `-Wincomplete-uni-patterns` as a part of `-Wall`, which
causes `Language.JVM.{CFG,Parser}` to produce warnings when built with 9.2. To
fix the warnings, I turned each partial uni-pattern into a `case` with an
explicit call to `error` if the pattern fails to match. This is no less partial
than before, but it avoids warnings and will produce a more specific error
message if the pattern fails to match.